Naming convention, Reduce Code
* Change all naming camel case

* Reduce code which is details of the read article page in 'BoardArticleDAO.java'

+++ src/main/webapp/BoardArticleDeleteForm.jsp
@@ -8,10 +8,10 @@
 	}
 	String num = request.getParameter("num");
 	BoardArticleDAO boardArticle = new BoardArticleDAO();
-	BoardArticleVO board = boardArticle.readArticle(num);	
-	if(Objects.equals(board.getWriter(),session.getAttribute("id"))){
+	BoardArticleVO board = boardArticle.readArticle(num);
+	String root = boardArticle.getRoot();
+	if(!Objects.equals(board.getWriter(),session.getAttribute("id"))&&!Objects.equals(root,session.getAttribute("id"))){
 		boardArticle.deleteArticle(num);	
-	} else {
 		out.println("<script>alert('작성자 혹은 관리자가 아닙니다.'); history.back();</script>");
 	}
